+++ Taehyung/AndroidStudy/features/Delivery/src/androidTest/java/com/terry/delivery/data/remote/SearchApiAndroidTest.kt
@@ -1,10 +1,12 @@
-package com.terry.delivery.data.local.remote
+package com.terry.delivery.data.remote
 
 import androidx.test.filters.SmallTest
 import com.google.common.truth.Truth
 import com.terry.delivery.ImmediateSchedulerRuleAndroidTest
 import com.terry.delivery.data.remote.LoginApi
+import com.terry.delivery.data.remote.SearchApi
 import com.terry.delivery.data.remote.model.LoginInfo
+import com.terry.delivery.util.Const
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.After
@@ -14,17 +16,20 @@ import org.junit.Test
 import javax.inject.Inject
 
 /*
- * Created by Taehyung Kim on 2021-10-14
+ * Created by Taehyung Kim on 2021-10-17
  */
 @SmallTest
 @HiltAndroidTest
-class LoginApiAndroidTest {
+class SearchApiAndroidTest {
 
     @get:Rule
     var hiltRule = HiltAndroidRule(this)
 
     @get:Rule
     var schedulerRule = ImmediateSchedulerRuleAndroidTest()
+
+    @Inject
+    lateinit var searchApi: SearchApi
 
     @Inject
     lateinit var loginApi: LoginApi
@@ -39,13 +44,19 @@ class LoginApiAndroidTest {
     }
 
     @Test
-    fun getAccessTokenWithValidInfo_returnsTrue() {
+    fun searchKeywordWithValidInfo_returnsTrue() {
         val loginInfo = LoginInfo("delivery", "dev_baemin")
-
         val token = loginApi.getAccessToken(loginInfo).blockingGet()
 
-        println("token -> ${token.access}")
-        Truth.assertThat(token.access).isNotEmpty()
-        Truth.assertThat(token.refresh).isNotEmpty()
+        val headers = Const.getSearchApiHeaders(token = token.access ?: "")
+
+        val result = searchApi.searchItem(
+            headerMap = headers,
+            query = "떡볶이",
+            page = 1
+        ).blockingGet()
+
+        Truth.assertThat(result.restaurantDto).isNotNull()
     }
+
 }
